Refactor to use `request` package for request handling.
Replaced all instances of `server.GetRequestBody` with `request.GetRequestBody` to standardize request body processing. Updated references and migrated `body.go` to the `request` package for better organization and modularity.
