+++ request/body.go
@@ -1,4 +1,4 @@
-package server
+package request
 
 import (
 	"encoding/json"
